+++ R/bundle.R
@@ -17,14 +17,14 @@ setMethod("names", "bundle", function(x) names(x@objects))
 #' @exportMethod names<-
 setReplaceMethod(
   "names",
-  signature = c(x = "bundle", value = "character"),
+  signature = c(x = "bundle", value = "vector"),
   function(x, value) {
+    if ( !is.vector(value) ){
+      stop("value needs to be a character vector")
+    }
+    if (is.list(value)) value <- unlist(value)
     if ( length(value) != length(x@objects) ) {
       warning("length of value provided does not match number of partitions")
-      stop()
-    }
-    if ( !is.character(names(x)) ){
-      warning("value needs to be a character vector")
       stop()
     }
     for (i in 1L:length(x@objects)) x@objects[[i]]@name <- value[i]