+++ NEWS.md
@@ -1,4 +1,5 @@
-# polmineR 0.7.9.9001
+# polmineR 0.7.10
+
 * The `use`-function is limited now to activating the corpus in data packages. Having
 introduced the session registry, switching registry directories is not needed any more.
 * The `partition_bundle`-method for `context`-objects has been reworked entirely (and is working again);
@@ -11,7 +12,8 @@ that is returned will include a report on weights.
 * The `as.regions`-function has been turned into a `as.regions`-method to have a more generic tool.
 * The `size_coi`-slot of the `context`-object included the node; the node (i.e. matches for queries) is excluded now from the count of the size_coi.
 * Some refactoring of the `context`-method, so that full use of data.table speeds up things.
-* 
+* `highlight`-method allows definitions of terms to be highlighted to be passed in via three dots (...);
+no explicit list necessary.
 
 # polmineR 0.7.9